modal_app/requirements.txt
modal>=0.60.0
firecrawl-py>=1.0.0
supabase>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sentry-sdk>=1.30.0
tenacity>=8.2.0
httpx>=0.25.0
python-dotenv>=1.0.0
